requirements.txt
beautifulsoup4==4.11.1
dnsdb==0.2.5
dnsdb_python==1.2.4
docx==0.2.4
Flask==2.0.2
Flask_Cors==3.0.10
networkx==2.6.3
oletools==0.60
OTXv2==1.5.12
pdfminer==20191125
pe==0.3.2
pefile==2021.9.3
py2neo==2021.2.3
pycryptodome==3.14.1
pyimpfuzzy==0.5
pymongo==4.0
python_docx==0.8.11
python_magic==0.4.25
r2pipe==1.6.5
requests==2.22.0
tldextract==2.2.2
urllib3==1.25.8
